Updated the dependency handling in the sprockets transformer
The previous version was building up a list of dependencies and passing
all of them to the compiler. I'm not sure why this was the case, as the
compiler only needs the entry point.

However, Sprockets does need to be made aware of these dependencies
so that it can watch those files and recompile on changes. This is
done with `depend_on`. I adapted this version to be more like the
Sprockets 2.x version in this gist https://gist.github.com/rtfeldman/db7b121100b6c6ff435b,
while maintaining the newer interface.