Remove test files from .gemspec
The tests now include two whole Rails applications, which has increased
the size of the packaged gem from 6kb to 1Mb. There is an open issue
on RubyGems to remove declared test_files from the packaged gem, but in
the meantime there does not appear to be any benefit in specifying them.